+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Added
 - Adds support to n (n > 0) http Seq consumer at Logger to prevent the application to consume a lot of memory when we have a thousands of requests
 - Change SeqClient.Send return type, from bool to error
+- Assert how many consumers was passed by user to the Logger
 
 ## 0.1.1 (2017-03-17)